Add common way to create Influx instance
This takes values from env variables and creates an InfluxDB object.
The influx object can then be passed to a tracker to track values.
